L3_Streamlit/textcomparison/requirements.txt
pdfminer.six==20231228
google-generativeai==0.7.0
streamlit==1.36.0